requirements.txt
autopep8==1.5.7
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
ecdsa==0.17.0
Flask==2.0.1
Flask-Login==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
python-jose==3.3.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
